Update PythonHTTRACK.py
Btw this is Caleb. This makes it so that it will not download any file more than 1mb as well as it allows it do download 8 files at a time so if that works we should have each thread take around 100 links. It takes a while to start up so we should try to minimize that by multi threading it inside the threads and allowing 100 at a time instead of 1 at a time.