Update PythonHTTRACK.py
Okay so the *c8 will allow you to access 8 different websites at a time and the -c16 allows 16 downloads on the same site. The reason that it was downloading everything was because we had the formatting on the %t flag. That is another thing I changed. -%s theoretically lowers bandwidth usage so that will be nice if I have to migrate it to my home server. The whole -F thing makes it so that it is not rejected by some sites because it is not identified as some mainstream browser.